Gracefully handle empty failure tags
This commit fixes #137. Some JUnit generators emit an empty failure tag,
with only a message property set. In those cases, the parser crashes
when trying to match the failure with a source file. Since this feature
is optional, the simplest fix is to skip the processing when the failure
tag is empty.

Also added a test, and the corresponding input file is generated from a
reporter within our codebase.